Add .spec files for HTTP components
This patch adds a "specs" directory containing spec files and build
scripts for packages that are needed by some oVirt components but that
aren't available in the supported distributions. Specifically it adds
specs for the following packages:

  httpcomponents-core
  httpcomponents-client

These two packages are needed by the oVirt Engine Java SDK, and they
aren't current available in the EL6 distributions or in the EPEL
repositories.

See the included README file for details on how to build the packages.

Change-Id: I5c01e9addea5fc7ba48683012b7704d1a2c91c35
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>