Remove the build script and .spec for libxml gem
A previous patch added a build script and .spec file to build the RPM
for the libxml Ruby gem, as that was needed for the ne Ruby SDK. But it
has been decided to avoid this dependency in the SDK, writing in C the
bridge between Ruby and libxml, as a Ruby extension instead. This patch
removes the script and the .spec completely.

Change-Id: I84ee0e13c069297154511c2da22b906610223555
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>