Added Node.js packages
This patch adds two .spec files needed to build Node.js and the NPM
modules that will be used by oVirt JavaScript applications like the
new dashboard that will be introduced in version 4.0 of the engine.

The first package "ovirt-engine-nodejs" is just a repackaging of the
Node.js binaries.

The second package "ovirt-engine-nodejs-modules" is the set of NPM
modules that will be needed to build the oVirt JavaScript applications.
At the moment the set of modules is composed of "less" and its
dependencies, and it is just an example.

Change-Id: I52d7680f9d6d9dc313640776311d59febb599ac3
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>