Move modules to a "node_modules" directory.
Currently this package installs all the modules inside
"/usr/share/ovirt-engine-nodejs-modules". But some tools doesn't like the
fact that the name of the directory isn't "node_modules". For example,
webpack will fail to find some of the modules it uses. Other tools may
fail in similar ways. To avoid this issue this patch moves all the
modules to a new "node_modules" directory inside
"ovirt-engine-nodejs-modules".

Change-Id: I9d701ad336e6c6cf74a721914251834893e43d9b
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>