Package tarball containing Node.js modules
Currently the "ovirt-engine-nodejs-modules" RPM package contains all the
required NPM modules required by oVirt projects. This means currently
more than 100000 files (yes, one hundred thousand) and that exceeds the
capacity of RPM headers. The right solution to that problem would be to
have each NPM package inside its own RPM package, but that is completely
unfeasible at this point. To avoid the RPM limit this patch changes the
"ovirt-engine-nodejs" so that it contains a tarball containing all the
files. This tarball is then expanded inside /var/lib. As users of the
package should use the "setup-env.sh" script to prepare their
environment, the won't notice the change, as this script takes care
of using the correct directory.

Change-Id: I3f3ca1af6a5602e451959d1771681d9dd10a7f5f
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>