Automatically RPM dependencies of Node.js modules
A previous patch changed the Node.js modules RPM so that it doesn't
contain directly the files of the modules, but a tarball. A side effect
of that is that RPM doesn't calculate automatically the list of
dependencies, as that calculation is based on the list of files directly
contined in the package. As a result some of the required libraries
(libfontconfig, for example) aren't installed when the package is
installed. To avoid that issue this patch changes the build script so
that it calculates the list of dependencies explicitly before generating
the RPM .spec file, using the same "find-requires" script used normally
by RPM.

Change-Id: Ie78263d61501cf1d5612708b4b8265e4b3a377d6
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>