release_notes_git: follow zstream clones, fix compare
Developers often forget to update Bug urls to the
zstream clones when cherry-picking to stable branches.
As a side effect, the fixed bugs are not added to the
release notes because of the wrong target milestone in
the referenced bug after the cloning.

This patch adress the issues by checking the bugs in the
block list searching for a valid clone with correct
target milestone.

In order to test this patch on 4.1.3 tc1 release
also moved from git bare repos to normal repos
in order to address errors while comparing git hashes
instead of tags.

Change-Id: I9906552f3d8032522fa3f52aaa331b2ab82ad8f2
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>