release_notes_git: Add authors translation table
It appears that some contributors are doing so from multiple
accounts, sometimes with their full name written in a different
way, i.e. with letters specific to their native language,
with dashes, with just the username etc. etc.

In order to properly count the individual contributors we had
to implement the translation table, so that contributions made
from any number of the related accounts will be properly connected
to the contributor, whose full and correct name will appear once.

This is the translation code implementation, as well as initial
translation table data, that fixes the issues that we noticed
for the current release.

Change-Id: I576eeff6c3db77f68eb8058581f5cc6b20ef65c2
Signed-off-by: Lev Veyde <lveyde@redhat.com>